+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.covidapp",
+    package = "dev.cloudgakkai.alfredo",
 )
 
 android_resource(
     name = "res",
-    package = "com.covidapp",
+    package = "dev.cloudgakkai.alfredo",
     res = "src/main/res",
 )
