+++ android/app/src/main/java/dev/cloudgakkai/alfredo/MainActivity.java
@@ -1,4 +1,4 @@
-package com.covidapp;
+package dev.cloudgakkai.alfredo;
 
 import com.facebook.react.ReactActivity;
 
@@ -24,6 +24,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "CovidApp";
+    return "Alfredo";
   }
 }
