+++ android/app/src/main/java/dev/cloudgakkai/alfredo/MainApplication.java
@@ -1,4 +1,4 @@
-package com.covidapp;
+package dev.cloudgakkai.alfredo;
 
 import android.app.Application;
 import android.content.Context;
@@ -66,7 +66,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.covidapp.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("dev.cloudgakkai.alfredo.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
