+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "CovidApp",
-  "displayName": "CovidApp"
+  "name": "Alfredo",
+  "displayName": "Alfredo"
 }
