+++ index.js
@@ -2,4 +2,4 @@ import './App/Config/ReactotronConfig'
 import { AppRegistry } from 'react-native'
 import App from './App/Containers/App'
 
-AppRegistry.registerComponent('CovidApp', () => App)
+AppRegistry.registerComponent('Alfredo', () => App)
