+++ ios/Alfredo/Images.xcassets/AppIcon.appiconset/Contents.json
