+++ ios/Alfredo/Images.xcassets/Contents.json
