+++ ios/AlfredoTests/AlfredoTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface CovidAppTests : XCTestCase
+@interface AlfredoTests : XCTestCase
 
 @end
 
-@implementation CovidAppTests
+@implementation AlfredoTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
