+++ ios/AlfredoTests/Info.plist
