+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "CovidApp",
+  "name": "Alfredo",
   "version": "0.0.1",
   "private": true,
   "scripts": {
