+++ storybook/storybook.js
@@ -9,5 +9,5 @@ configure(() => {
 // This assumes that storybook is running on the same host as your RN packager,
 // to set manually use, e.g. host: 'localhost' option
 const StorybookUI = getStorybookUI({ port: 7007, onDeviceUI: true })
-AppRegistry.registerComponent('CovidApp', () => StorybookUI)
+AppRegistry.registerComponent('Alfredo', () => StorybookUI)
 export default StorybookUI
